sample size calculation handles nans/infs (#42)
* sample size calculation handles nans/infs

* nice assert

+++ src/epstats/server/res.py
@@ -280,7 +280,7 @@ class SampleSizeCalculationResult(BaseModel):
     Result of the sample size calculation.
     """
 
-    sample_size_per_variant: int = Field(
+    sample_size_per_variant: float = Field(
         title="Sample size per variant",
     )
